Creates common VenueForm for edit and create
